modularização da função de uso do kms
